Chore 식별자 Long으로 변경

+++ src/main/java/escaper/backend/repository/cafe/CafeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface CafeRepository extends JpaRepository<Cafe, Integer>, CafeRepositoryCustom {
+public interface CafeRepository extends JpaRepository<Cafe, Long>, CafeRepositoryCustom {
 
     @Query("SELECT DISTINCT c.address.city FROM Cafe c")
     List<String> searchCities();
